Moved the tests under the main namespace and autoload them.
Closes #478
