+++ tests/Exception/ExpectationExceptionTest.php
@@ -1,11 +1,18 @@
 <?php
 
-namespace Tests\Behat\Mink\Exception;
+namespace Behat\Mink\Tests\Exception;
 
-use Behat\Mink\Exception\ResponseTextException;
+use Behat\Mink\Exception\ExpectationException;
 
-class ResponseTextExceptionTest extends \PHPUnit_Framework_TestCase
+class ExpectationExceptionTest extends \PHPUnit_Framework_TestCase
 {
+    public function testEmptyMessageAndPreviousException()
+    {
+        $exception = new ExpectationException('', $this->getSessionMock(), new \Exception('Something failed'));
+
+        $this->assertEquals('Something failed', $exception->getMessage());
+    }
+
     public function testExceptionToString()
     {
         $driver = $this->getMock('Behat\Mink\Driver\DriverInterface');
@@ -26,24 +33,38 @@ class ResponseTextExceptionTest extends \PHPUnit_Framework_TestCase
             ->will($this->returnValue('http://localhost/test'));
 
         $page->expects($this->any())
-            ->method('getText')
-            ->will($this->returnValue("Hello world\nTest\n"));
+            ->method('getContent')
+            ->will($this->returnValue("<html><head>\n<title>Hello</title>\n</head>\n<body>\n<h1>Hello world</h1>\n<p>Test</p>\n</body></html>"));
 
         $expected = <<<'TXT'
-Text error
+Expectation failure
 
 +--[ HTTP/1.1 200 | http://localhost/test | %s ]
 |
-|  Hello world
-|  Test
+|  <body>
+|  <h1>Hello world</h1>
+|  <p>Test</p>
+|  </body>
 |
 TXT;
 
         $expected = sprintf($expected.'  ', get_class($driver));
 
-        $exception = new ResponseTextException('Text error', $session);
+        $exception = new ExpectationException('Expectation failure', $session);
 
         $this->assertEquals($expected, $exception->__toString());
+    }
+
+    public function testExceptionWhileRenderingString()
+    {
+        $session = $this->getSessionMock();
+        $session->expects($this->any())
+            ->method('getPage')
+            ->will($this->throwException(new \Exception('Broken page')));
+
+        $exception = new ExpectationException('Expectation failure', $session);
+
+        $this->assertEquals('Expectation failure', $exception->__toString());
     }
 
     private function getSessionMock()
