+++ tests/Selector/PartialNamedSelectorTest.php
@@ -1,14 +1,9 @@
 <?php
 
-namespace Tests\Behat\Mink\Selector;
+namespace Behat\Mink\Tests\Selector;
 
 use Behat\Mink\Selector\PartialNamedSelector;
 
-require_once __DIR__ . '/NamedSelectorTest.php';
-
-/**
- * @group unittest
- */
 class PartialNamedSelectorTest extends NamedSelectorTest
 {
     protected function getSelector()
